# swagger/Trainingsplanner-postgres.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "laplanner",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "tags": [
          "TrainingsGroup"
        ],
        "operationId": "TrainingsGroup_GetAllGroups",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsGroupDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TrainingsGroup"
        ],
        "operationId": "TrainingsGroup_CreateGroup",
        "requestBody": {
          "x-name": "trainingsGroupDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsGroupDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsGroupDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TrainingsGroup"
        ],
        "operationId": "TrainingsGroup_UpdateGroup",
        "requestBody": {
          "x-name": "trainingsGroupDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsGroupDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsGroupDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TrainingsGroup"
        ],
        "operationId": "TrainingsGroup_DeleteGroup",
        "requestBody": {
          "x-name": "trainingsGroupDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsGroupDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{id}": {
      "get": {
        "tags": [
          "TrainingsGroup"
        ],
        "operationId": "TrainingsGroup_GetGroupById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsGroupDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{id}/appointments": {
      "get": {
        "tags": [
          "TrainingsGroup"
        ],
        "operationId": "TrainingsGroup_GetAppointmentsByGroupId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsAppointmentDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groupusers/athlete": {
      "post": {
        "tags": [
          "TrainingsGroupUser"
        ],
        "operationId": "TrainingsGroupUser_AddUserToGroup",
        "requestBody": {
          "x-name": "trainingsGroupApplicationUserDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsGroupApplicationUserDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groupusers/trainer": {
      "post": {
        "tags": [
          "TrainingsGroupUser"
        ],
        "operationId": "TrainingsGroupUser_AddTrainerToGroup",
        "requestBody": {
          "x-name": "trainingsGroupApplicationUserDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsGroupApplicationUserDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groupusers": {
      "delete": {
        "tags": [
          "TrainingsGroupUser"
        ],
        "operationId": "TrainingsGroupUser_DeleteMemberFromGroup",
        "parameters": [
          {
            "name": "trainingsGroupId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "userId",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TrainingsAppointmentDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          }
        }
      },
      "TrainingsGroup": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "trainingsGroupsApplicationUsers": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsGroupApplicationUser"
            }
          },
          "trainingsAppointments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointment"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsGroupApplicationUser": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "applicationUserId": {
            "type": "string",
            "nullable": true
          },
          "isTrainer": {
            "type": "boolean"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsAppointment": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsAppointmentTrainingsModule": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsAppointmentId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsAppointment": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsAppointment"
              }
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsModule": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "$ref": "#/components/schemas/TrainingsDifficulty"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "trainingsModulesTrainingsExercises": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsExercise"
            }
          },
          "trainingsModulesTrainingsModuleTags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsModuleTag"
            }
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsDifficulty": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "UNKNOWN",
          "EASY",
          "MEDIUM",
          "HARD",
          "IMPOSSIBLE"
        ],
        "enum": [
          0,
          10,
          20,
          30,
          40
        ]
      },
      "TrainingsModuleTrainingsExercise": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsExerciesId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsExercise": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsExercise"
              }
            ]
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsExercise": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "repetitions": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModulesTrainingsExercises": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsExercise"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsModuleTrainingsModuleTag": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsModuleTagId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModuleTag": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModuleTag"
              }
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsModuleTag": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "trainingsModulesTrainingsModuleTags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsModuleTag"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TrainingsGroupDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "trainingsGroupsApplicationUsers": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsGroupApplicationUser"
            }
          },
          "trainingsAppointments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointment"
            }
          }
        }
      },
      "TrainingsGroupApplicationUserDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "applicationUserId": {
            "type": "string",
            "nullable": true
          },
          "isTrainer": {
            "type": "boolean"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  }
}